+++ PersistencePlugin/src/main/java/com/elmakers/mine/bukkit/persistence/annotation/Migrate.java
@@ -1,5 +1,6 @@
 package com.elmakers.mine.bukkit.persistence.annotation;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -18,11 +19,10 @@ import java.lang.annotation.Target;
  * @see PersistClass
  *
  */
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface Migrate
 {
-	boolean autoReset() default false;
-	
 	MigrateStep[] steps() default {};
 }