+++ PersistencePlugin/src/main/java/com/elmakers/mine/bukkit/persistence/annotation/PersistClass.java
@@ -1,5 +1,6 @@
 package com.elmakers.mine.bukkit.persistence.annotation;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -17,6 +18,8 @@ import java.lang.annotation.Target;
  * @see PersistField
  *
  */
+
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface PersistClass